Add --error flag to kubectl port-forward
This new flag controls how port forwarding handles
errors that occur.

If the flag is set to "exit", when an error occurs
while forwarding, the listener will be closed and
the command will exit. This is the current behavior
and the default value for this new flag.

If the flag is set to "continue", when an error
occurs while forwarding, the listener will not be
closed. This is equivalent to the port forwarding
behavior prior to kubectl 1.23, which some users
rely upon in certain use cases.