Fix bug in update-translations.sh escape sequence handling
Fixed a bug in hack/update-translations.sh, where it would not
correctly double backslashes in the original text, resulting
in an "invalid control sequence" error when run with the -x
flag to update template.pot.

After the fix, I ran hack/update-translations.sh -x to update
template.pot with the latest translatable text.