Leave connection open when error occurs on handled port forwarding connection.
When an error occurs in a handled connection, the overall port forwarding session should not be ended because we don't know if the error is fatal or not.

In the case of some errors, like "connection reset by peer", it is not fatal and closing the entire port forwarding session makes certain applications unusable with port forwarding as connection resets often happen during the normal course of operation.